+++ CHANGELOG.rst
@@ -17,6 +17,7 @@ Removed
 ^^^^^^^
 * Dropped python 3.7 support
 * Disallowed parsing cookbooks of a later major version
+* Dropped implicit targets
 
 `4.4.0`_ - 2023-01-31
 ---------------------